--- administration/src/models/CardCreationModel.ts
@@ -1,9 +1,0 @@
-import {CardType} from "./CardType";
-
-export interface CardCreationModel {
-    id: number,
-    forename: string,
-    surname: string,
-    expirationDate: Date,
-    cardType: CardType
-}
